move service_settings into subfolder
it's a pretty big file. we already have a subfolder for testing so lets
make a subfolder for the endpoints too
