+++ tests/app/main/views/service_settings/test_service_settings.py
@@ -14,7 +14,7 @@ from notifications_python_client.errors import HTTPError
 from notifications_utils.clients.zendesk.zendesk_client import NotifySupportTicket
 
 import app
-from app.main.views.service_settings import _should_set_default_org_branding
+from app.main.views.service_settings.index import _should_set_default_org_branding
 from app.models.service import Service
 from tests import (
     find_element_by_tag_and_partial_text,
@@ -847,7 +847,7 @@ def test_should_check_for_reply_to_on_go_live(
     )
 
     mock_get_reply_to_email_addresses = mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_reply_to_email_addresses",
+        "app.main.views.service_settings.index.service_api_client.get_reply_to_email_addresses",
         return_value=reply_to_email_addresses,
     )
 
@@ -1138,7 +1138,7 @@ def test_should_check_for_sms_sender_on_go_live(
     )
 
     mock_get_sms_senders = mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_sms_senders",
+        "app.main.views.service_settings.index.service_api_client.get_sms_senders",
         return_value=sms_senders,
     )
 
@@ -1192,11 +1192,11 @@ def test_should_check_for_mou_on_request_to_go_live(
         return_value=[],
     )
     mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_sms_senders",
+        "app.main.views.service_settings.index.service_api_client.get_sms_senders",
         return_value=[],
     )
     mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_reply_to_email_addresses",
+        "app.main.views.service_settings.index.service_api_client.get_reply_to_email_addresses",
         return_value=[],
     )
     for channel in {"email", "sms", "letter"}:
@@ -1240,11 +1240,11 @@ def test_gp_without_organisation_is_shown_agreement_step(
         return_value=[],
     )
     mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_sms_senders",
+        "app.main.views.service_settings.index.service_api_client.get_sms_senders",
         return_value=[],
     )
     mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_reply_to_email_addresses",
+        "app.main.views.service_settings.index.service_api_client.get_reply_to_email_addresses",
         return_value=[],
     )
     for channel in {"email", "sms", "letter"}:
@@ -1290,11 +1290,11 @@ def test_non_gov_user_is_told_they_cant_go_live(
         return_value=[],
     )
     mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_sms_senders",
+        "app.main.views.service_settings.index.service_api_client.get_sms_senders",
         return_value=[],
     )
     mocker.patch(
-        "app.main.views.service_settings.service_api_client.get_reply_to_email_addresses",
+        "app.main.views.service_settings.index.service_api_client.get_reply_to_email_addresses",
         return_value=[],
     )
     client_request.login(api_nongov_user_active)
@@ -1566,7 +1566,7 @@ def test_should_redirect_after_request_to_go_live(
         )
     mock_create_ticket = mocker.spy(NotifySupportTicket, "__init__")
     mock_send_ticket_to_zendesk = mocker.patch(
-        "app.main.views.service_settings.zendesk_client.send_ticket_to_zendesk",
+        "app.main.views.service_settings.index.zendesk_client.send_ticket_to_zendesk",
         autospec=True,
     )
     page = client_request.post("main.request_to_go_live", service_id=SERVICE_ONE_ID, _follow_redirects=True)
@@ -1640,7 +1640,7 @@ def test_request_to_go_live_displays_go_live_notes_in_zendesk_ticket(
     )
     mock_create_ticket = mocker.spy(NotifySupportTicket, "__init__")
     mock_send_ticket_to_zendesk = mocker.patch(
-        "app.main.views.service_settings.zendesk_client.send_ticket_to_zendesk",
+        "app.main.views.service_settings.index.zendesk_client.send_ticket_to_zendesk",
         autospec=True,
     )
     client_request.post("main.request_to_go_live", service_id=SERVICE_ONE_ID, _follow_redirects=True)
@@ -1711,7 +1711,7 @@ def test_request_to_go_live_displays_mou_signatories(
         ),
     )
     mocker.patch(
-        "app.main.views.service_settings.zendesk_client.send_ticket_to_zendesk",
+        "app.main.views.service_settings.index.zendesk_client.send_ticket_to_zendesk",
         autospec=True,
     )
     mock_create_ticket = mocker.spy(NotifySupportTicket, "__init__")
@@ -1747,7 +1747,9 @@ def test_should_be_able_to_request_to_go_live_with_no_organisation(
             new_callable=PropertyMock,
             return_value=1,
         )
-    mock_post = mocker.patch("app.main.views.service_settings.zendesk_client.send_ticket_to_zendesk", autospec=True)
+    mock_post = mocker.patch(
+        "app.main.views.service_settings.index.zendesk_client.send_ticket_to_zendesk", autospec=True
+    )
 
     client_request.post("main.request_to_go_live", service_id=SERVICE_ONE_ID, _follow_redirects=True)
 
@@ -3911,10 +3913,10 @@ def test_POST_email_branding_enter_government_identity_logo_text(
     mocker, client_request, service_one, extra_url_args, expected_ticket_content, expected_extra_url_args
 ):
     mock_send_ticket_to_zendesk = mocker.patch(
-        "app.main.views.service_settings.zendesk_client.send_ticket_to_zendesk",
+        "app.main.views.service_settings.index.zendesk_client.send_ticket_to_zendesk",
         autospec=True,
     )
-    mock_flash = mocker.patch("app.main.views.service_settings.flash", autospec=True)
+    mock_flash = mocker.patch("app.main.views.service_settings.index.flash", autospec=True)
 
     client_request.post(
         "main.email_branding_enter_government_identity_logo_text",
@@ -4091,9 +4093,9 @@ def test_GET_email_branding_upload_logo(
 )
 def test_POST_email_branding_upload_logo_success(mocker, client_request, service_one, email_branding_data):
     antivirus_mock = mocker.patch("app.extensions.antivirus_client.scan", return_value=True)
-    mock_upload_email_logo = mocker.patch("app.main.views.service_settings.upload_email_logo")
+    mock_upload_email_logo = mocker.patch("app.main.views.service_settings.index.upload_email_logo")
     mock_upload_email_logo.return_value = "my-logo-path"
-    mocker.patch("app.main.views.service_settings.uuid.uuid4", return_value="my-logo-uuid")
+    mocker.patch("app.main.views.service_settings.index.uuid.uuid4", return_value="my-logo-uuid")
 
     mocker.patch.dict(
         "flask.current_app.config", {"EMAIL_BRANDING_MIN_LOGO_HEIGHT_PX": 1, "EMAIL_BRANDING_MAX_LOGO_WIDTH_PX": 1}
@@ -4152,7 +4154,7 @@ def test_POST_email_branding_upload_logo_validation_errors(
     if callable(post_data):
         post_data = post_data()
 
-    mock_upload_email_logo = mocker.patch("app.main.views.service_settings.upload_email_logo")
+    mock_upload_email_logo = mocker.patch("app.main.views.service_settings.index.upload_email_logo")
 
     with mock.patch.dict("app.main.validators.current_app.config", {"ANTIVIRUS_ENABLED": False}):
         page = client_request.post(
@@ -4177,7 +4179,7 @@ def test_POST_email_branding_upload_logo_validation_errors(
 def test_POST_email_branding_upload_logo_enforces_minimum_logo_height(
     mocker, client_request, service_one, min_logo_height, expect_error
 ):
-    mocker.patch("app.main.views.service_settings.upload_email_logo")
+    mocker.patch("app.main.views.service_settings.index.upload_email_logo")
     mocker.patch("app.utils.image_processing.ImageProcessor")
 
     with mock.patch.dict(
@@ -4200,7 +4202,7 @@ def test_POST_email_branding_upload_logo_enforces_minimum_logo_height(
 
 
 def test_POST_email_branding_upload_logo_resizes_and_pads_wide_short_logo(mocker, client_request, service_one):
-    mocker.patch("app.main.views.service_settings.upload_email_logo")
+    mocker.patch("app.main.views.service_settings.index.upload_email_logo")
     mock_image_processor = mocker.patch("app.main.forms.ImageProcessor")
     mock_image_processor().height = ComparablePropertyMock(side_effect=[26, 13])
     mock_image_processor().width = 100
@@ -4327,9 +4329,9 @@ def test_POST_email_branding_set_alt_text_creates_branding_adds_to_pool_and_redi
     brand_type,
     expected_name,
 ):
-    mock_flash = mocker.patch("app.main.views.service_settings.flash")
+    mock_flash = mocker.patch("app.main.views.service_settings.index.flash")
     mock_should_set_default_org_branding = mocker.patch(
-        "app.main.views.service_settings._should_set_default_org_branding", return_value=False
+        "app.main.views.service_settings.index._should_set_default_org_branding", return_value=False
     )
     mock_add_to_branding_pool = mocker.patch(
         "app.organisations_client.add_brandings_to_email_branding_pool", return_value=None
@@ -4379,7 +4381,7 @@ def test_POST_email_branding_set_alt_text_creates_branding_sets_org_default_if_a
 ):
     service_one["organisation"] = ORGANISATION_ID
     mock_should_set_default_org_branding = mocker.patch(
-        "app.main.views.service_settings._should_set_default_org_branding", return_value=True
+        "app.main.views.service_settings.index._should_set_default_org_branding", return_value=True
     )
     mock_add_to_branding_pool = mocker.patch(
         "app.organisations_client.add_brandings_to_email_branding_pool", return_value=None
@@ -4861,7 +4863,7 @@ def test_archive_service_after_confirm(
 ):
     service_one["restricted"] = is_trial_service
     mock_api = mocker.patch("app.service_api_client.post")
-    mock_event = mocker.patch("app.main.views.service_settings.create_archive_service_event")
+    mock_event = mocker.patch("app.main.views.service_settings.index.create_archive_service_event")
     redis_delete_mock = mocker.patch("app.notify_client.service_api_client.redis_client.delete")
     mocker.patch("app.notify_client.service_api_client.redis_client.delete_by_pattern")
 
@@ -5250,7 +5252,7 @@ def test_post_service_receive_text_messages_start_turns_on_feature_and_redirects
         "app.inbound_number_client.add_inbound_number_to_service",
         return_value={"id": "abcd", "service_id": SERVICE_ONE_ID, "inbound_number_id": "1234"},
     )
-    mock_event = mocker.patch("app.main.views.service_settings.create_set_inbound_sms_on_event")
+    mock_event = mocker.patch("app.main.views.service_settings.index.create_set_inbound_sms_on_event")
 
     page = client_request.post(
         ".service_receive_text_messages_start", service_id=SERVICE_ONE_ID, _follow_redirects=True
@@ -6050,7 +6052,9 @@ def test_service_confirm_broadcast_account_type_posts_data_to_api_and_redirects(
     mock_get_users_by_service,
 ):
     set_service_broadcast_settings_mock = mocker.patch("app.service_api_client.set_service_broadcast_settings")
-    mock_event_handler = mocker.patch("app.main.views.service_settings.create_broadcast_account_type_change_event")
+    mock_event_handler = mocker.patch(
+        "app.main.views.service_settings.index.create_broadcast_account_type_change_event"
+    )
 
     client_request.login(platform_admin_user)
     client_request.post(
@@ -6086,7 +6090,9 @@ def test_service_confirm_broadcast_account_type_errors_for_unknown_type(
     account_type,
 ):
     set_service_broadcast_settings_mock = mocker.patch("app.service_api_client.set_service_broadcast_settings")
-    mock_event_handler = mocker.patch("app.main.views.service_settings.create_broadcast_account_type_change_event")
+    mock_event_handler = mocker.patch(
+        "app.main.views.service_settings.index.create_broadcast_account_type_change_event"
+    )
 
     client_request.login(platform_admin_user)
     client_request.post(
